Updates epochs to be parsed using a regular expression.
Instead of using NimbleParsec to parse epoch timestamps, we will now use
a regular expression.

This change should allow for better handling of parsing subseconds in an
epoch. When doing this parsing with
[`NimpleParsec.integer/2`](https://hexdocs.pm/nimble_parsec/NimbleParsec.html#integer/2)
leading `0`s would be dropped. However, when working with subseconds in
an epoch time stamp, these `0` are important to have.

When deciding how to fix this issue and also thinking about how to add
support for negative epochs, it seemed like `NimbleParsec` was making
things more complicated than necessary. As a result, it was decided
using a regular expression should work well and be easier to change.