Fix timezone conversion to UTC (#20)
Resolves #18

We were using the current timestamp (essentially DateTime.utc_now()) when converting to UTC, which uses the UTC offset of today instead of the UTC offset in the context of the parsed datetime.

In other words, `to_utc: true` would always convert using now's UTC offset instead of the parsed datetime's. So if today was Daylight Time, and the parsed datetime was in Standard Time, the parsed datetime would be converted to UTC using the daylight time UTC offset. This resulted in incorrect conversion.