Fixes subsecond handling for epoch parsing.
Fixes how subseconds/microseconds are handled for epoch timestamps when
the subseconds are preceded by zeros.

Previously, we would parse the subseconds as an Integer, and lose
preceding zeros. This commit changes the logic to:

1. Parse the subseconds as a Float to keep leading zeros
2. "Convert" subseconds into microseconds by multiplying them by 10^6
  - We truncate this number to drop anything after the decimal point
  - We chose truncate over round because naive rounding can result in
adding a second to the timestamp
3. Create a tuple that is compatible with sending the microseconds on a
`DateTime`, that is something in the shape of `{microsecond, length}`.
We use `6` as an upper bound because that is the most subsecond digits
support by `DateTime`.
  - If there are no subseconds, we return `{0,0}`, which doesn't end up
adding anything to a `DateTime`.

Now, `Epoch.parse` is responsible for converting the seconds and
subseconds into something more easily compatible with `DateTime`, and
`Epoch.from_tokens` has less code; most of it was moved into
`Epoch.parse` because we needed to know the length of the subsecond
string that was parsed.