Convert epoch parsing to use regular expressions and better handle retraining subsecond precision  (#16)
* Updates epochs to be parsed using a regular expression.

Instead of using NimbleParsec to parse epoch timestamps, we will now use
a regular expression.

This change should allow for better handling of parsing subseconds in an
epoch. When doing this parsing with
[`NimpleParsec.integer/2`](https://hexdocs.pm/nimble_parsec/NimbleParsec.html#integer/2)
leading `0`s would be dropped. However, when working with subseconds in
an epoch time stamp, these `0` are important to have.

When deciding how to fix this issue and also thinking about how to add
support for negative epochs, it seemed like `NimbleParsec` was making
things more complicated than necessary. As a result, it was decided
using a regular expression should work well and be easier to change.

* Removes unused `Combinators.Epoch`.
* Fixes subsecond handling for epoch parsing.

Fixes how subseconds/microseconds are handled for epoch timestamps when
the subseconds are preceded by zeros.

Previously, we would parse the subseconds as an Integer, and lose
preceding zeros. This commit changes the logic to:

1. Parse the subseconds as a Float to keep leading zeros
2. "Convert" subseconds into microseconds by multiplying them by 10^6
  - We truncate this number to drop anything after the decimal point
  - We chose truncate over round because naive rounding can result in
adding a second to the timestamp
3. Create a tuple that is compatible with sending the microseconds on a
`DateTime`, that is something in the shape of `{microsecond, length}`.
We use `6` as an upper bound because that is the most subsecond digits
support by `DateTime`.
  - If there are no subseconds, we return `{0,0}`, which doesn't end up
adding anything to a `DateTime`.

Now, `Epoch.parse` is responsible for converting the seconds and
subseconds into something more easily compatible with `DateTime`, and
`Epoch.from_tokens` has less code; most of it was moved into
`Epoch.parse` because we needed to know the length of the subsecond
string that was parsed.

* Makes `Epoch.parse` more restrictive.

Previously, when parsing the parsed "seconds" portion of an epoch
timestamp, we ignore any extra characters. To be safe, we now expect
there to be **no** extra characters.