Adds support for parsing negative epoch timestamps.
When checking if a given date is in the epoch format, `epoch_regex` will
now also match if there is a `-`, meaning a negative timestamp. From the
[Wikipedia article](https://en.wikipedia.org/wiki/Unix_time):

> This can be extended backwards from the epoch too, using negative
> numbers; thus 1957-10-04T00:00:00Z, 4472 days before the epoch, is
> represented by the Unix time number −4472 × 86400 = −386380800. This
> applies within days as well; the time number at any given time of a day
> is the number of seconds that has passed since the midnight starting
> that day added to the time number of that midnight.

If there are no subseconds, we can simply multiple the number of seconds
by -1, and let `DateTime.from_unix` will convert correctly.

If there are subseconds, we need to handle the difference - we
substract an additional seconds since we are moving backwards in time.
We then calculate the difference between the additional second we moved
back and the number of microseconds given (1 second = 10e6 microseconds
- the number of subseconds in the timestamp).