Fixes issue with negative epoch with subseconds.
Fixes an issue where a negative epoch could have its subseconds
truncated down to 0. Previously, we would always do `1e6 -
truncated_subseconds`, however, if the subseconds are truncated down to
0, this left us with `1e6`. Now, we check if the subseconds have been
truncated down to `0` and skip the subtraction in that case.

This issue came up with writing tests, so this commit also includes
various other tests.