Adds support for parsing negative epoch timestamps. … (#24)
* Adds support for parsing negative epoch timestamps.

When checking if a given date is in the epoch format, `epoch_regex` will
now also match if there is a `-`, meaning a negative timestamp. From the
[Wikipedia article](https://en.wikipedia.org/wiki/Unix_time):

> This can be extended backwards from the epoch too, using negative
> numbers; thus 1957-10-04T00:00:00Z, 4472 days before the epoch, is
> represented by the Unix time number −4472 × 86400 = −386380800. This
> applies within days as well; the time number at any given time of a day
> is the number of seconds that has passed since the midnight starting
> that day added to the time number of that midnight.

If there are no subseconds, we can simply multiple the number of seconds
by -1, and let `DateTime.from_unix` will convert correctly.

If there are subseconds, we need to handle the difference - we
substract an additional seconds since we are moving backwards in time.
We then calculate the difference between the additional second we moved
back and the number of microseconds given (1 second = 10e6 microseconds
- the number of subseconds in the timestamp).

* Refactors `Epoch.parse`.

Extracts the functionality for parsing seconds and subseconds into their
own functions.

* Fixes issue with negative epoch with subseconds.

Fixes an issue where a negative epoch could have its subseconds
truncated down to 0. Previously, we would always do `1e6 -
truncated_subseconds`, however, if the subseconds are truncated down to
0, this left us with `1e6`. Now, we check if the subseconds have been
truncated down to `0` and skip the subtraction in that case.

This issue came up with writing tests, so this commit also includes
various other tests.

* Refactors `Epoch.parse_subseconds/2`.

Creates a module attribute `one_second_in_microseconds`. This should
have two advantages:

* Hopefully, the name helps make it clearer what is going on
* The calculation can be done once at compile time instead of every time
the function is run

* Updates docs to include negative epoch support.

* Updates README and `DateTimeParser` module docs to update the epoch
section to include negative epoch support
* Adds doc tests to `DateTimeParser` to include negative epochs
* Also, adds doc tests that include subseconds